fix(kernel): decouple cache guard and classify worker panics
Tracking: #1710
