+++ changes/1710.security.md
@@ -16,15 +16,19 @@
   principal runtime scratch; lockless injected kernels preserve another live
   kernel's roots. Tracking #1710.
 - Retained admitted callback filesystem jobs across callback-connection
-  cancellation. A drain timeout or filesystem-worker join failure reports
-  cleanup failure at `drain` and leaves the exact revoked UUID root, provider
-  resources, and mutation blocker mapped; exact retry completes only after
-  jobs and the fence finish. Tracking #1710.
+  cancellation. A drain timeout, filesystem-worker join failure, or inner
+  `spawn_blocking` panic reports cleanup failure at `drain` and leaves the
+  exact revoked UUID root, provider resources, and mutation blocker mapped;
+  exact retry completes only after jobs and the fence finish. Tracking #1710.
 - Closed the retain interval after initial projection validation but before
   reference acquisition: authorized revocation or expiry performs exact
   cleanup and admits a replacement in the same mount request, without acquiring
   a stale reference. Provider drift or incomplete cleanup remains an
   administrative blocker. Tracking #1710.
+- Separated projection startup serialization from the projection-cache guard
+  so retain can never re-enter that guard while mount owns it. This preserves
+  singleton provider startup while first-mount revoke and expiry intervals
+  exact-clean and admit their replacement without deadlock. Tracking #1710.
 - Removed the public raw process-storage remover. The kernel-private validated
   cleanup is admitted only by the existing private native singleton gate;
   arbitrary `Some(File)` resources cannot manufacture boot ownership.
